Define InMemoryStorage in matter.idl.generators.storage
